# [Enhancement] check mem limit after fetch one chunk from storage engine