## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #4711 

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

Currently we cannot `TryCatch` the memory allocation of the storage layer, so we need to add MemCheck.

Every time a Chunk is taken out from the storage engine, a memory check will be executed.

I have test it in benchmark env, don't affect the performance of large scan.

16core 64G mem

cpu usage 1550%

before modify

```
-concurrency=1 --number-of-queries=10 --create-schema=ssb --query="select max(s_address) from lineorder_flat where split(s_city, ' ')[1] in ('CHINA', 'CANADA', 'INDIA');"
Benchmark
        Average number of seconds to run all queries: 25.861 seconds
        Minimum number of seconds to run all queries: 25.861 seconds
        Maximum number of seconds to run all queries: 25.861 seconds
        Number of clients running queries: 1
        Average number of queries per client: 10

-concurrency=5 --number-of-queries=40 --create-schema=ssb --query="select max(s_address) from lineorder_flat where split(s_city, ' ')[1] in ('CHINA', 'CANADA', 'INDIA');"
mysqlslap: [Warning] Using a password on the command line interface can be insecure.
Benchmark
        Average number of seconds to run all queries: 94.693 seconds
        Minimum number of seconds to run all queries: 94.693 seconds
        Maximum number of seconds to run all queries: 94.693 seconds
        Number of clients running queries: 5
        Average number of queries per client: 8

-concurrency=10 --number-of-queries=50 --create-schema=ssb --query="select max(s_address) from lineorder_flat where split(s_city, ' ')[1] in ('CHINA', 'CANADA', 'INDIA');"
mysqlslap: [Warning] Using a password on the command line interface can be insecure.
Benchmark
        Average number of seconds to run all queries: 117.633 seconds
        Minimum number of seconds to run all queries: 117.633 seconds
        Maximum number of seconds to run all queries: 117.633 seconds
        Number of clients running queries: 10
        Average number of queries per client: 5
```

after modify 

```
--concurrency=1 --number-of-queries=10 --create-schema=ssb --query="select max(s_address) from lineorder_flat where split(s_city, ' ')[1] in ('CHINA', 'CANADA', 'INDIA');"
Benchmark
        Average number of seconds to run all queries: 25.746 seconds
        Minimum number of seconds to run all queries: 25.746 seconds
        Maximum number of seconds to run all queries: 25.746 seconds
        Number of clients running queries: 1
        Average number of queries per client: 10


-concurrency=5 --number-of-queries=40 --create-schema=ssb --query="select max(s_ad                                                                                                  
dress) from lineorder_flat where split(s_city, ' ')[1] in ('CHINA', 'CANADA', 'INDIA');"                                                                                                                                                                                                                                                                                                                                                                                    
Benchmark                                                                                                                                                                                                                                                                         
        Average number of seconds to run all queries: 94.235 seconds                                                                                                                                                                                                              
        Minimum number of seconds to run all queries: 94.235 seconds                                                                                                                                                                                                              
        Maximum number of seconds to run all queries: 94.235 seconds                                                                     
        Number of clients running queries: 5                                                                                             
        Average number of queries per client: 8

-concurrency=10 --number-of-queries=50 --create-schema=ssb --query="select max(s_address) from lineorder_flat where split(s_city, ' ')[1] in ('CHINA', 'CANADA', 'INDIA');"                                                
Benchmark                                                           
        Average number of seconds to run all queries: 117.665 seconds                                                                                                                                                                                                             
        Minimum number of seconds to run all queries: 117.665 seconds                                                                                                                                                                                                             
        Maximum number of seconds to run all queries: 117.665 seconds                                                                                                                                                                                                             
        Number of clients running queries: 10                                                                                            
        Average number of queries per client: 5
```

Through perf top, you can't see the performance cost of this code